requirements.txt
numpy==1.17.2
opencv_python==4.1.1.26
seaborn==0.9.0
tensorflow-gpu==1.14.0
Keras==2.2.5
matplotlib==3.1.1
pandas==0.25.1
scikit_learn==0.21.3